void static variable for bge_objects duplis list
